+++ txscript/engine_test.go
@@ -1,15 +1,14 @@
 // Copyright (c) 2013-2016 The btcsuite developers
-// Copyright (c) 2015-2016 The Decred developers
+// Copyright (c) 2015-2018 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package txscript_test
+package txscript
 
 import (
 	"testing"
 
 	"github.com/decred/dcrd/chaincfg/chainhash"
-	"github.com/decred/dcrd/txscript"
 	"github.com/decred/dcrd/wire"
 )
 
@@ -18,60 +17,50 @@ import (
 func TestBadPC(t *testing.T) {
 	t.Parallel()
 
-	type pcTest struct {
+	tests := []struct {
 		script, off int
-	}
-	pcTests := []pcTest{
-		{
-			script: 2,
-			off:    0,
-		},
-		{
-			script: 0,
-			off:    2,
-		},
+	}{
+		{script: 2, off: 0},
+		{script: 0, off: 2},
 	}
 	// tx with almost empty scripts.
 	tx := &wire.MsgTx{
 		SerType: wire.TxSerializeFull,
 		Version: 1,
-		TxIn: []*wire.TxIn{
-			{
-				PreviousOutPoint: wire.OutPoint{
-					Hash: chainhash.Hash([32]byte{
-						0xc9, 0x97, 0xa5, 0xe5,
-						0x6e, 0x10, 0x41, 0x02,
-						0xfa, 0x20, 0x9c, 0x6a,
-						0x85, 0x2d, 0xd9, 0x06,
-						0x60, 0xa2, 0x0b, 0x2d,
-						0x9c, 0x35, 0x24, 0x23,
-						0xed, 0xce, 0x25, 0x85,
-						0x7f, 0xcd, 0x37, 0x04,
-					}),
-					Index: 0,
-				},
-				SignatureScript: []uint8{txscript.OP_NOP},
-				Sequence:        4294967295,
+		TxIn: []*wire.TxIn{{
+			PreviousOutPoint: wire.OutPoint{
+				Hash: chainhash.Hash([32]byte{
+					0xc9, 0x97, 0xa5, 0xe5,
+					0x6e, 0x10, 0x41, 0x02,
+					0xfa, 0x20, 0x9c, 0x6a,
+					0x85, 0x2d, 0xd9, 0x06,
+					0x60, 0xa2, 0x0b, 0x2d,
+					0x9c, 0x35, 0x24, 0x23,
+					0xed, 0xce, 0x25, 0x85,
+					0x7f, 0xcd, 0x37, 0x04,
+				}),
+				Index: 0,
 			},
-		},
-		TxOut: []*wire.TxOut{
-			{
-				Value:    1000000000,
-				PkScript: nil,
-			},
-		},
+			SignatureScript: mustParseShortForm("NOP"),
+			Sequence:        4294967295,
+		}},
+		TxOut: []*wire.TxOut{{
+			Value:    1000000000,
+			PkScript: nil,
+		}},
 		LockTime: 0,
 	}
-	pkScript := []byte{txscript.OP_NOP}
+	pkScript := mustParseShortForm("NOP")
 
-	for _, test := range pcTests {
-		vm, err := txscript.NewEngine(pkScript, tx, 0, 0, 0, nil)
+	for _, test := range tests {
+		vm, err := NewEngine(pkScript, tx, 0, 0, 0, nil)
 		if err != nil {
 			t.Errorf("Failed to create script: %v", err)
 		}
 
 		// set to after all scripts
-		vm.TstSetPC(test.script, test.off)
+		vm.scriptIdx = test.script
+		vm.scriptOff = test.off
 
 		_, err = vm.Step()
 		if err == nil {
@@ -123,21 +112,10 @@ func TestCheckErrorCondition(t *testing.T) {
 		},
 		LockTime: 0,
 	}
-	pkScript := []byte{
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_NOP,
-		txscript.OP_TRUE,
-	}
+	pkScript := mustParseShortForm("NOP NOP NOP NOP NOP NOP NOP NOP NOP" +
+		" NOP TRUE")
 
-	vm, err := txscript.NewEngine(pkScript, tx, 0, 0, 0, nil)
+	vm, err := NewEngine(pkScript, tx, 0, 0, 0, nil)
 	if err != nil {
 		t.Errorf("failed to create script: %v", err)
 	}
@@ -154,7 +132,7 @@ func TestCheckErrorCondition(t *testing.T) {
 		}
 
 		err = vm.CheckErrorCondition(false)
-		if err != txscript.ErrStackScriptUnfinished {
+		if err != ErrStackScriptUnfinished {
 			t.Errorf("got unexepected error %v on %dth iteration",
 				err, i)
 			return
@@ -181,46 +159,42 @@ func TestCheckErrorCondition(t *testing.T) {
 func TestInvalidFlagCombinations(t *testing.T) {
 	t.Parallel()
 
-	tests := []txscript.ScriptFlags{
-		txscript.ScriptVerifyCleanStack,
+	tests := []ScriptFlags{
+		ScriptVerifyCleanStack,
 	}
 
 	// tx with almost empty scripts.
 	tx := &wire.MsgTx{
 		SerType: wire.TxSerializeFull,
 		Version: 1,
-		TxIn: []*wire.TxIn{
-			{
-				PreviousOutPoint: wire.OutPoint{
-					Hash: chainhash.Hash([32]byte{
-						0xc9, 0x97, 0xa5, 0xe5,
-						0x6e, 0x10, 0x41, 0x02,
-						0xfa, 0x20, 0x9c, 0x6a,
-						0x85, 0x2d, 0xd9, 0x06,
-						0x60, 0xa2, 0x0b, 0x2d,
-						0x9c, 0x35, 0x24, 0x23,
-						0xed, 0xce, 0x25, 0x85,
-						0x7f, 0xcd, 0x37, 0x04,
-					}),
-					Index: 0,
-				},
-				SignatureScript: []uint8{txscript.OP_NOP},
-				Sequence:        4294967295,
+		TxIn: []*wire.TxIn{{
+			PreviousOutPoint: wire.OutPoint{
+				Hash: chainhash.Hash([32]byte{
+					0xc9, 0x97, 0xa5, 0xe5,
+					0x6e, 0x10, 0x41, 0x02,
+					0xfa, 0x20, 0x9c, 0x6a,
+					0x85, 0x2d, 0xd9, 0x06,
+					0x60, 0xa2, 0x0b, 0x2d,
+					0x9c, 0x35, 0x24, 0x23,
+					0xed, 0xce, 0x25, 0x85,
+					0x7f, 0xcd, 0x37, 0x04,
+				}),
+				Index: 0,
 			},
-		},
-		TxOut: []*wire.TxOut{
-			{
-				Value:    1000000000,
-				PkScript: nil,
-			},
-		},
+			SignatureScript: []uint8{OP_NOP},
+			Sequence:        4294967295,
+		}},
+		TxOut: []*wire.TxOut{{
+			Value:    1000000000,
+			PkScript: nil,
+		}},
 		LockTime: 0,
 	}
-	pkScript := []byte{txscript.OP_NOP}
+	pkScript := []byte{OP_NOP}
 
 	for i, test := range tests {
-		_, err := txscript.NewEngine(pkScript, tx, 0, test, 0, nil)
-		if err != txscript.ErrInvalidFlags {
+		_, err := NewEngine(pkScript, tx, 0, test, 0, nil)
+		if err != ErrInvalidFlags {
 			t.Fatalf("TestInvalidFlagCombinations #%d unexpected "+
 				"error: %v", i, err)
 		}
@@ -239,7 +213,7 @@ func TestCheckPubKeyEncoding(t *testing.T) {
 	}{
 		{
 			name: "uncompressed ok",
-			key: decodeHex("0411db93e1dcdb8a016b49840f8c53bc1eb68" +
+			key: hexToBytes("0411db93e1dcdb8a016b49840f8c53bc1eb68" +
 				"a382e97b1482ecad7b148a6909a5cb2e0eaddfb84ccf" +
 				"9744464f82e160bfa9b8b64f9d4c03f999b8643f656b" +
 				"412a3"),
@@ -247,19 +221,19 @@ func TestCheckPubKeyEncoding(t *testing.T) {
 		},
 		{
 			name: "compressed ok",
-			key: decodeHex("02ce0b14fb842b1ba549fdd675c98075f12e9" +
+			key: hexToBytes("02ce0b14fb842b1ba549fdd675c98075f12e9" +
 				"c510f8ef52bd021a9a1f4809d3b4d"),
 			isValid: true,
 		},
 		{
 			name: "compressed ok",
-			key: decodeHex("032689c7c2dab13309fb143e0e8fe39634252" +
+			key: hexToBytes("032689c7c2dab13309fb143e0e8fe39634252" +
 				"1887e976690b6b47f5b2a4b7d448e"),
 			isValid: true,
 		},
 		{
 			name: "hybrid",
-			key: decodeHex("0679be667ef9dcbbac55a06295ce870b07029" +
+			key: hexToBytes("0679be667ef9dcbbac55a06295ce870b07029" +
 				"bfcdb2dce28d959f2815b16f81798483ada7726a3c46" +
 				"55da4fbfc0e1108a8fd17b448a68554199c47d08ffb1" +
 				"0d4b8"),
@@ -272,15 +246,14 @@ func TestCheckPubKeyEncoding(t *testing.T) {
 		},
 	}
 
-	flags := txscript.ScriptVerifyStrictEncoding
+	vm := Engine{flags: ScriptVerifyStrictEncoding}
 	for _, test := range tests {
-		err := txscript.TstCheckPubKeyEncoding(test.key, flags)
+		err := vm.checkPubKeyEncoding(test.key)
 		if err != nil && test.isValid {
-			t.Errorf("checkSignatureEncoding test '%s' failed "+
-				"when it should have succeeded: %v", test.name,
-				err)
+			t.Errorf("checkPubKeyEncoding test '%s' failed when "+
+				"it should have succeeded: %v", test.name, err)
 		} else if err == nil && !test.isValid {
-			t.Errorf("checkSignatureEncooding test '%s' succeeded "+
+			t.Errorf("checkPubKeyEncoding test '%s' succeeded "+
 				"when it should have failed", test.name)
 		}
 	}
@@ -299,7 +272,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 	}{
 		{
 			name: "valid signature",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -312,7 +285,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "bad magic",
-			sig: decodeHex("314402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("314402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -320,7 +293,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "bad 1st int marker magic",
-			sig: decodeHex("304403204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304403204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -328,7 +301,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "bad 2nd int marker",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41032018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -336,7 +309,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "short len",
-			sig: decodeHex("304302204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304302204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -344,7 +317,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "long len",
-			sig: decodeHex("304502204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304502204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -352,7 +325,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "long X",
-			sig: decodeHex("304402424e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402424e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -360,7 +333,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "long Y",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022118152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -368,7 +341,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "short Y",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41021918152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -376,7 +349,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "trailing crap",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d0901"),
@@ -384,7 +357,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "X == N ",
-			sig: decodeHex("30440220fffffffffffffffffffffffffffff" +
+			sig: hexToBytes("30440220fffffffffffffffffffffffffffff" +
 				"ffebaaedce6af48a03bbfd25e8cd0364141022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -392,7 +365,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "X == N ",
-			sig: decodeHex("30440220fffffffffffffffffffffffffffff" +
+			sig: hexToBytes("30440220fffffffffffffffffffffffffffff" +
 				"ffebaaedce6af48a03bbfd25e8cd0364142022018152" +
 				"2ec8eca07de4860a4acdd12909d831cc56cbbac46220" +
 				"82221a8768d1d09"),
@@ -400,7 +373,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "Y == N",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd410220fffff" +
 				"ffffffffffffffffffffffffffebaaedce6af48a03bb" +
 				"fd25e8cd0364141"),
@@ -408,7 +381,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "Y > N",
-			sig: decodeHex("304402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd410220fffff" +
 				"ffffffffffffffffffffffffffebaaedce6af48a03bb" +
 				"fd25e8cd0364142"),
@@ -416,19 +389,19 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "0 len X",
-			sig: decodeHex("302402000220181522ec8eca07de4860a4acd" +
+			sig: hexToBytes("302402000220181522ec8eca07de4860a4acd" +
 				"d12909d831cc56cbbac4622082221a8768d1d09"),
 			isValid: false,
 		},
 		{
 			name: "0 len Y",
-			sig: decodeHex("302402204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("302402204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd410200"),
 			isValid: false,
 		},
 		{
 			name: "extra R padding",
-			sig: decodeHex("30450221004e45e16932b8af514961a1d3a1a" +
+			sig: hexToBytes("30450221004e45e16932b8af514961a1d3a1a" +
 				"25fdf3f4f7732e9d624c6c61548ab5fb8cd410220181" +
 				"522ec8eca07de4860a4acdd12909d831cc56cbbac462" +
 				"2082221a8768d1d09"),
@@ -436,7 +409,7 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 		{
 			name: "extra S padding",
-			sig: decodeHex("304502204e45e16932b8af514961a1d3a1a25" +
+			sig: hexToBytes("304502204e45e16932b8af514961a1d3a1a25" +
 				"fdf3f4f7732e9d624c6c61548ab5fb8cd41022100181" +
 				"522ec8eca07de4860a4acdd12909d831cc56cbbac462" +
 				"2082221a8768d1d09"),
@@ -444,9 +417,9 @@ func TestCheckSignatureEncoding(t *testing.T) {
 		},
 	}
 
-	flags := txscript.ScriptVerifyStrictEncoding
+	vm := Engine{flags: ScriptVerifyStrictEncoding}
 	for _, test := range tests {
-		err := txscript.TstCheckSignatureEncoding(test.sig, flags)
+		err := vm.checkSignatureEncoding(test.sig)
 		if err != nil && test.isValid {
 			t.Errorf("checkSignatureEncoding test '%s' failed "+
 				"when it should have succeeded: %v", test.name,
