+++ txscript/standard_test.go
@@ -1,46 +1,19 @@
 // Copyright (c) 2013-2015 The btcsuite developers
-// Copyright (c) 2015-2016 The Decred developers
+// Copyright (c) 2015-2018 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package txscript_test
+package txscript
 
 import (
 	"bytes"
-	"encoding/hex"
 	"reflect"
 	"testing"
 
 	"github.com/decred/dcrd/chaincfg"
 	"github.com/decred/dcrd/chaincfg/chainec"
 	"github.com/decred/dcrd/dcrutil"
-	"github.com/decred/dcrd/txscript"
 )
-
-// decodeHex decodes the passed hex string and returns the resulting bytes.  It
-// panics if an error occurs.  This is only used in the tests as a helper since
-// the only way it can fail is if there is an error in the test source code.
-func decodeHex(hexStr string) []byte {
-	b, err := hex.DecodeString(hexStr)
-	if err != nil {
-		panic("invalid hex string in test source: err " + err.Error() +
-			", hex: " + hexStr)
-	}
-
-	return b
-}
-
-// hexToBytes converts the passed hex string into bytes and will panic if there
-// is an error.  This is only provided for the hard-coded constants so errors in
-// the source code can be detected. It will only (and must only) be called with
-// hard-coded values.
-func hexToBytes(s string) []byte {
-	b, err := hex.DecodeString(s)
-	if err != nil {
-		panic("invalid hex in source file: " + s)
-	}
-	return b
-}
 
 // mustParseShortForm parses the passed short form script and returns the
 // resulting bytes.  It panics if an error occurs.  This is only used in the
@@ -80,7 +53,7 @@ func newAddressPubKey(serializedPubKey []byte) dcrutil.Address {
 // test source code.
 func newAddressPubKeyHash(pkHash []byte) dcrutil.Address {
 	addr, err := dcrutil.NewAddressPubKeyHash(pkHash, &chaincfg.MainNetParams,
-		secp)
+		chainec.ECTypeSecp256k1)
 	if err != nil {
 		panic("invalid public key hash in test source")
 	}
@@ -112,144 +85,144 @@ func TestExtractPkScriptAddrs(t *testing.T) {
 		script  []byte
 		addrs   []dcrutil.Address
 		reqSigs int
-		class   txscript.ScriptClass
+		class   ScriptClass
 		noparse bool
 	}{
 		{
 			name: "standard p2pk with compressed pubkey (0x02)",
-			script: decodeHex("2102192d74d0cb94344c9569c2e7790157" +
-				"3d8d7903c3ebec3a957724895dca52c6b4ac"),
+			script: hexToBytes("2102192d74d0cb94344c9569c2e779015" +
+				"73d8d7903c3ebec3a957724895dca52c6b4ac"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("02192d74d0cb94344" +
-					"c9569c2e77901573d8d7903c3ebec3a95772" +
-					"4895dca52c6b4")),
+				newAddressPubKey(hexToBytes("02192d74d0cb9434" +
+					"4c9569c2e77901573d8d7903c3ebec3a9577" +
+					"24895dca52c6b4")),
 			},
 			reqSigs: 1,
-			class:   txscript.PubKeyTy,
+			class:   PubKeyTy,
 		},
 		{
 			name: "standard p2pk with uncompressed pubkey (0x04)",
-			script: decodeHex("410411db93e1dcdb8a016b49840f8c53bc" +
-				"1eb68a382e97b1482ecad7b148a6909a5cb2e0eaddfb" +
-				"84ccf9744464f82e160bfa9b8b64f9d4c03f999b8643" +
-				"f656b412a3ac"),
+			script: hexToBytes("410411db93e1dcdb8a016b49840f8c53b" +
+				"c1eb68a382e97b1482ecad7b148a6909a5cb2e0eaddf" +
+				"b84ccf9744464f82e160bfa9b8b64f9d4c03f999b864" +
+				"3f656b412a3ac"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("0411db93e1dcdb8a0" +
-					"16b49840f8c53bc1eb68a382e97b1482ecad" +
-					"7b148a6909a5cb2e0eaddfb84ccf9744464f" +
-					"82e160bfa9b8b64f9d4c03f999b8643f656b" +
-					"412a3")),
+				newAddressPubKey(hexToBytes("0411db93e1dcdb8a" +
+					"016b49840f8c53bc1eb68a382e97b1482eca" +
+					"d7b148a6909a5cb2e0eaddfb84ccf9744464" +
+					"f82e160bfa9b8b64f9d4c03f999b8643f656" +
+					"b412a3")),
 			},
 			reqSigs: 1,
-			class:   txscript.PubKeyTy,
+			class:   PubKeyTy,
 		},
 		{
 			name: "standard p2pk with compressed pubkey (0x03)",
-			script: decodeHex("2103b0bd634234abbb1ba1e986e884185c" +
-				"61cf43e001f9137f23c2c409273eb16e65ac"),
+			script: hexToBytes("2103b0bd634234abbb1ba1e986e884185" +
+				"c61cf43e001f9137f23c2c409273eb16e65ac"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("03b0bd634234abbb1" +
-					"ba1e986e884185c61cf43e001f9137f23c2c" +
-					"409273eb16e65")),
+				newAddressPubKey(hexToBytes("03b0bd634234abbb" +
+					"1ba1e986e884185c61cf43e001f9137f23c2" +
+					"c409273eb16e65")),
 			},
 			reqSigs: 1,
-			class:   txscript.PubKeyTy,
+			class:   PubKeyTy,
 		},
 		{
 			name: "2nd standard p2pk with uncompressed pubkey (0x04)",
-			script: decodeHex("4104b0bd634234abbb1ba1e986e884185c" +
-				"61cf43e001f9137f23c2c409273eb16e6537a576782e" +
-				"ba668a7ef8bd3b3cfb1edb7117ab65129b8a2e681f3c" +
-				"1e0908ef7bac"),
+			script: hexToBytes("4104b0bd634234abbb1ba1e986e884185" +
+				"c61cf43e001f9137f23c2c409273eb16e6537a576782" +
+				"eba668a7ef8bd3b3cfb1edb7117ab65129b8a2e681f3" +
+				"c1e0908ef7bac"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("04b0bd634234abbb1" +
-					"ba1e986e884185c61cf43e001f9137f23c2c" +
-					"409273eb16e6537a576782eba668a7ef8bd3" +
-					"b3cfb1edb7117ab65129b8a2e681f3c1e090" +
-					"8ef7b")),
+				newAddressPubKey(hexToBytes("04b0bd634234abbb" +
+					"1ba1e986e884185c61cf43e001f9137f23c2" +
+					"c409273eb16e6537a576782eba668a7ef8bd" +
+					"3b3cfb1edb7117ab65129b8a2e681f3c1e09" +
+					"08ef7b")),
 			},
 			reqSigs: 1,
-			class:   txscript.PubKeyTy,
+			class:   PubKeyTy,
 		},
 		{
 			name: "standard p2pkh",
-			script: decodeHex("76a914ad06dd6ddee55cbca9a9e3713bd7" +
-				"587509a3056488ac"),
+			script: hexToBytes("76a914ad06dd6ddee55cbca9a9e3713bd" +
+				"7587509a3056488ac"),
 			addrs: []dcrutil.Address{
-				newAddressPubKeyHash(decodeHex("ad06dd6ddee55" +
-					"cbca9a9e3713bd7587509a30564")),
+				newAddressPubKeyHash(hexToBytes("ad06dd6ddee5" +
+					"5cbca9a9e3713bd7587509a30564")),
 			},
 			reqSigs: 1,
-			class:   txscript.PubKeyHashTy,
+			class:   PubKeyHashTy,
 		},
 		{
 			name: "standard p2sh",
-			script: decodeHex("a91463bcc565f9e68ee0189dd5cc67f1b0" +
-				"e5f02f45cb87"),
+			script: hexToBytes("a91463bcc565f9e68ee0189dd5cc67f1b" +
+				"0e5f02f45cb87"),
 			addrs: []dcrutil.Address{
-				newAddressScriptHash(decodeHex("63bcc565f9e68" +
-					"ee0189dd5cc67f1b0e5f02f45cb")),
+				newAddressScriptHash(hexToBytes("63bcc565f9e6" +
+					"8ee0189dd5cc67f1b0e5f02f45cb")),
 			},
 			reqSigs: 1,
-			class:   txscript.ScriptHashTy,
+			class:   ScriptHashTy,
 		},
 		// from real tx 60a20bd93aa49ab4b28d514ec10b06e1829ce6818ec06cd3aabd013ebcdc4bb1, vout 0
 		{
 			name: "standard 1 of 2 multisig",
-			script: decodeHex("514104cc71eb30d653c0c3163990c47b97" +
-				"6f3fb3f37cccdcbedb169a1dfef58bbfbfaff7d8a473" +
-				"e7e2e6d317b87bafe8bde97e3cf8f065dec022b51d11" +
-				"fcdd0d348ac4410461cbdcc5409fb4b4d42b51d33381" +
-				"354d80e550078cb532a34bfa2fcfdeb7d76519aecc62" +
-				"770f5b0e4ef8551946d8a540911abe3e7854a26f39f5" +
-				"8b25c15342af52ae"),
+			script: hexToBytes("514104cc71eb30d653c0c3163990c47b9" +
+				"76f3fb3f37cccdcbedb169a1dfef58bbfbfaff7d8a47" +
+				"3e7e2e6d317b87bafe8bde97e3cf8f065dec022b51d1" +
+				"1fcdd0d348ac4410461cbdcc5409fb4b4d42b51d3338" +
+				"1354d80e550078cb532a34bfa2fcfdeb7d76519aecc6" +
+				"2770f5b0e4ef8551946d8a540911abe3e7854a26f39f" +
+				"58b25c15342af52ae"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("04cc71eb30d653c0c" +
-					"3163990c47b976f3fb3f37cccdcbedb169a1" +
-					"dfef58bbfbfaff7d8a473e7e2e6d317b87ba" +
-					"fe8bde97e3cf8f065dec022b51d11fcdd0d3" +
-					"48ac4")),
-				newAddressPubKey(decodeHex("0461cbdcc5409fb4b" +
-					"4d42b51d33381354d80e550078cb532a34bf" +
-					"a2fcfdeb7d76519aecc62770f5b0e4ef8551" +
-					"946d8a540911abe3e7854a26f39f58b25c15" +
-					"342af")),
+				newAddressPubKey(hexToBytes("04cc71eb30d653c0" +
+					"c3163990c47b976f3fb3f37cccdcbedb169a" +
+					"1dfef58bbfbfaff7d8a473e7e2e6d317b87b" +
+					"afe8bde97e3cf8f065dec022b51d11fcdd0d" +
+					"348ac4")),
+				newAddressPubKey(hexToBytes("0461cbdcc5409fb4" +
+					"b4d42b51d33381354d80e550078cb532a34b" +
+					"fa2fcfdeb7d76519aecc62770f5b0e4ef855" +
+					"1946d8a540911abe3e7854a26f39f58b25c1" +
+					"5342af")),
 			},
 			reqSigs: 1,
-			class:   txscript.MultiSigTy,
+			class:   MultiSigTy,
 		},
 		// from real tx d646f82bd5fbdb94a36872ce460f97662b80c3050ad3209bef9d1e398ea277ab, vin 1
 		{
 			name: "standard 2 of 3 multisig",
-			script: decodeHex("524104cb9c3c222c5f7a7d3b9bd152f363" +
-				"a0b6d54c9eb312c4d4f9af1e8551b6c421a6a4ab0e29" +
-				"105f24de20ff463c1c91fcf3bf662cdde4783d4799f7" +
-				"87cb7c08869b4104ccc588420deeebea22a7e900cc8b" +
-				"68620d2212c374604e3487ca08f1ff3ae12bdc639514" +
-				"d0ec8612a2d3c519f084d9a00cbbe3b53d071e9b09e7" +
-				"1e610b036aa24104ab47ad1939edcb3db65f7fedea62" +
-				"bbf781c5410d3f22a7a3a56ffefb2238af8627363bdf" +
-				"2ed97c1f89784a1aecdb43384f11d2acc64443c7fc29" +
-				"9cef0400421a53ae"),
+			script: hexToBytes("524104cb9c3c222c5f7a7d3b9bd152f36" +
+				"3a0b6d54c9eb312c4d4f9af1e8551b6c421a6a4ab0e2" +
+				"9105f24de20ff463c1c91fcf3bf662cdde4783d4799f" +
+				"787cb7c08869b4104ccc588420deeebea22a7e900cc8" +
+				"b68620d2212c374604e3487ca08f1ff3ae12bdc63951" +
+				"4d0ec8612a2d3c519f084d9a00cbbe3b53d071e9b09e" +
+				"71e610b036aa24104ab47ad1939edcb3db65f7fedea6" +
+				"2bbf781c5410d3f22a7a3a56ffefb2238af8627363bd" +
+				"f2ed97c1f89784a1aecdb43384f11d2acc64443c7fc2" +
+				"99cef0400421a53ae"),
 			addrs: []dcrutil.Address{
-				newAddressPubKey(decodeHex("04cb9c3c222c5f7a7" +
-					"d3b9bd152f363a0b6d54c9eb312c4d4f9af1" +
-					"e8551b6c421a6a4ab0e29105f24de20ff463" +
-					"c1c91fcf3bf662cdde4783d4799f787cb7c0" +
-					"8869b")),
-				newAddressPubKey(decodeHex("04ccc588420deeebe" +
-					"a22a7e900cc8b68620d2212c374604e3487c" +
-					"a08f1ff3ae12bdc639514d0ec8612a2d3c51" +
-					"9f084d9a00cbbe3b53d071e9b09e71e610b0" +
-					"36aa2")),
-				newAddressPubKey(decodeHex("04ab47ad1939edcb3" +
-					"db65f7fedea62bbf781c5410d3f22a7a3a56" +
-					"ffefb2238af8627363bdf2ed97c1f89784a1" +
-					"aecdb43384f11d2acc64443c7fc299cef040" +
-					"0421a")),
+				newAddressPubKey(hexToBytes("04cb9c3c222c5f7a" +
+					"7d3b9bd152f363a0b6d54c9eb312c4d4f9af" +
+					"1e8551b6c421a6a4ab0e29105f24de20ff46" +
+					"3c1c91fcf3bf662cdde4783d4799f787cb7c" +
+					"08869b")),
+				newAddressPubKey(hexToBytes("04ccc588420deeeb" +
+					"ea22a7e900cc8b68620d2212c374604e3487" +
+					"ca08f1ff3ae12bdc639514d0ec8612a2d3c5" +
+					"19f084d9a00cbbe3b53d071e9b09e71e610b" +
+					"036aa2")),
+				newAddressPubKey(hexToBytes("04ab47ad1939edcb" +
+					"3db65f7fedea62bbf781c5410d3f22a7a3a5" +
+					"6ffefb2238af8627363bdf2ed97c1f89784a" +
+					"1aecdb43384f11d2acc64443c7fc299cef04" +
+					"00421a")),
 			},
 			reqSigs: 2,
-			class:   txscript.MultiSigTy,
+			class:   MultiSigTy,
 		},
 
 		// The below are nonstandard script due to things such as
@@ -258,23 +231,23 @@ func TestExtractPkScriptAddrs(t *testing.T) {
 
 		{
 			name: "p2pk with uncompressed pk missing OP_CHECKSIG",
-			script: decodeHex("410411db93e1dcdb8a016b49840f8c53bc" +
-				"1eb68a382e97b1482ecad7b148a6909a5cb2e0eaddfb" +
-				"84ccf9744464f82e160bfa9b8b64f9d4c03f999b8643" +
-				"f656b412a3"),
+			script: hexToBytes("410411db93e1dcdb8a016b49840f8c53b" +
+				"c1eb68a382e97b1482ecad7b148a6909a5cb2e0eaddf" +
+				"b84ccf9744464f82e160bfa9b8b64f9d4c03f999b864" +
+				"3f656b412a3"),
 			addrs:   nil,
 			reqSigs: 0,
-			class:   txscript.NonStandardTy,
+			class:   NonStandardTy,
 		},
 		{
 			name: "valid signature from a sigscript - no addresses",
-			script: decodeHex("47304402204e45e16932b8af514961a1d3" +
-				"a1a25fdf3f4f7732e9d624c6c61548ab5fb8cd410220" +
-				"181522ec8eca07de4860a4acdd12909d831cc56cbbac" +
-				"4622082221a8768d1d0901"),
+			script: hexToBytes("47304402204e45e16932b8af514961a1d" +
+				"3a1a25fdf3f4f7732e9d624c6c61548ab5fb8cd41022" +
+				"0181522ec8eca07de4860a4acdd12909d831cc56cbba" +
+				"c4622082221a8768d1d0901"),
 			addrs:   nil,
 			reqSigs: 0,
-			class:   txscript.NonStandardTy,
+			class:   NonStandardTy,
 		},
 		// Note the technically the pubkey is the second item on the
 		// stack, but since the address extraction intentionally only
@@ -282,72 +255,72 @@ func TestExtractPkScriptAddrs(t *testing.T) {
 		// addresses.
 		{
 			name: "valid sigscript to reedeem p2pk - no addresses",
-			script: decodeHex("493046022100ddc69738bf2336318e4e04" +
-				"1a5a77f305da87428ab1606f023260017854350ddc02" +
-				"2100817af09d2eec36862d16009852b7e3a0f6dd7659" +
-				"8290b7834e1453660367e07a014104cd4240c198e125" +
-				"23b6f9cb9f5bed06de1ba37e96a1bbd13745fcf9d11c" +
-				"25b1dff9a519675d198804ba9962d3eca2d5937d58e5" +
-				"a75a71042d40388a4d307f887d"),
+			script: hexToBytes("493046022100ddc69738bf2336318e4e0" +
+				"41a5a77f305da87428ab1606f023260017854350ddc0" +
+				"22100817af09d2eec36862d16009852b7e3a0f6dd765" +
+				"98290b7834e1453660367e07a014104cd4240c198e12" +
+				"523b6f9cb9f5bed06de1ba37e96a1bbd13745fcf9d11" +
+				"c25b1dff9a519675d198804ba9962d3eca2d5937d58e" +
+				"5a75a71042d40388a4d307f887d"),
 			addrs:   nil,
 			reqSigs: 0,
-			class:   txscript.NonStandardTy,
+			class:   NonStandardTy,
 		},
 		// from real tx 691dd277dc0e90a462a3d652a1171686de49cf19067cd33c7df0392833fb986a, vout 0
 		// invalid public keys
 		{
 			name: "1 of 3 multisig with invalid pubkeys",
-			script: decodeHex("51411c2200007353455857696b696c6561" +
-				"6b73204361626c6567617465204261636b75700a0a63" +
-				"61626c65676174652d3230313031323034313831312e" +
-				"377a0a0a446f41776e6c6f61642074686520666f6c6c" +
-				"6f77696e67207472616e73616374696f6e7320776974" +
-				"68205361746f736869204e616b616d6f746f27732064" +
-				"6f776e6c6f61416420746f6f6c2077686963680a6361" +
-				"6e20626520666f756e6420696e207472616e73616374" +
-				"696f6e20366335336364393837313139656637393764" +
-				"35616463636453ae"),
+			script: hexToBytes("51411c2200007353455857696b696c656" +
+				"16b73204361626c6567617465204261636b75700a0a6" +
+				"361626c65676174652d3230313031323034313831312" +
+				"e377a0a0a446f41776e6c6f61642074686520666f6c6" +
+				"c6f77696e67207472616e73616374696f6e732077697" +
+				"468205361746f736869204e616b616d6f746f2773206" +
+				"46f776e6c6f61416420746f6f6c2077686963680a636" +
+				"16e20626520666f756e6420696e207472616e7361637" +
+				"4696f6e2036633533636439383731313965663739376" +
+				"435616463636453ae"),
 			addrs:   []dcrutil.Address{},
 			reqSigs: 1,
-			class:   txscript.MultiSigTy,
+			class:   MultiSigTy,
 		},
 		// from real tx: 691dd277dc0e90a462a3d652a1171686de49cf19067cd33c7df0392833fb986a, vout 44
 		// invalid public keys
 		{
 			name: "1 of 3 multisig with invalid pubkeys 2",
-			script: decodeHex("5141346333656332353963373464616365" +
-				"36666430383862343463656638630a63363662633139" +
-				"39366338623934613338313162333635363138666531" +
-				"65396231623541366361636365393933613339383861" +
-				"34363966636336643664616266640a32363633636661" +
-				"39636634633033633630396335393363336539316665" +
-				"64653730323921313233646434326432353633396433" +
-				"38613663663530616234636434340a00000053ae"),
+			script: hexToBytes("514134633365633235396337346461636" +
+				"536666430383862343463656638630a6336366263313" +
+				"93936633862393461333831316233363536313866653" +
+				"16539623162354136636163636539393361333938386" +
+				"134363966636336643664616266640a3236363363666" +
+				"13963663463303363363039633539336333653931666" +
+				"56465373032392131323364643432643235363339643" +
+				"338613663663530616234636434340a00000053ae"),
 			addrs:   []dcrutil.Address{},
 			reqSigs: 1,
-			class:   txscript.MultiSigTy,
+			class:   MultiSigTy,
 		},
 		{
 			name:    "empty script",
 			script:  []byte{},
 			addrs:   nil,
 			reqSigs: 0,
-			class:   txscript.NonStandardTy,
+			class:   NonStandardTy,
 		},
 		{
 			name:    "script that does not parse",
-			script:  []byte{txscript.OP_DATA_45},
+			script:  []byte{OP_DATA_45},
 			addrs:   nil,
 			reqSigs: 0,
-			class:   txscript.NonStandardTy,
+			class:   NonStandardTy,
 			noparse: true,
 		},
 	}
 
 	t.Logf("Running %d tests.", len(tests))
 	for i, test := range tests {
-		class, addrs, reqSigs, err := txscript.ExtractPkScriptAddrs(
-			txscript.DefaultScriptVersion, test.script,
+		class, addrs, reqSigs, err := ExtractPkScriptAddrs(
+			DefaultScriptVersion, test.script,
 			&chaincfg.MainNetParams)
 		if err != nil && !test.noparse {
 			t.Errorf("ExtractPkScriptAddrs #%d (%s): %v", i,
@@ -387,7 +360,7 @@ func TestCalcScriptInfo(t *testing.T) {
 		sigScript     string
 		pkScript      string
 		bip16         bool
-		scriptInfo    txscript.ScriptInfo
+		scriptInfo    ScriptInfo
 		scriptInfoErr error
 	}{
 		{
@@ -399,7 +372,7 @@ func TestCalcScriptInfo(t *testing.T) {
 			pkScript: "HASH160 DATA_20 0xfe441065b6532231de2fac56" +
 				"3152205ec4f59c",
 			bip16:         true,
-			scriptInfoErr: txscript.ErrStackShortScript,
+			scriptInfoErr: ErrStackShortScript,
 		},
 		{
 			name: "sigScript doesn't parse",
@@ -409,7 +382,7 @@ func TestCalcScriptInfo(t *testing.T) {
 			pkScript: "HASH160 DATA_20 0xfe441065b6532231de2fac56" +
 				"3152205ec4f59c74 EQUAL",
 			bip16:         true,
-			scriptInfoErr: txscript.ErrStackShortScript,
+			scriptInfoErr: ErrStackShortScript,
 		},
 		{
 			// Invented scripts, the hashes do not match
@@ -420,8 +393,8 @@ func TestCalcScriptInfo(t *testing.T) {
 			pkScript: "HASH160 DATA_20 0xfe441065b6532231de2fac56" +
 				"3152205ec4f59c74 EQUAL",
 			bip16: true,
-			scriptInfo: txscript.ScriptInfo{
-				PkScriptClass:  txscript.ScriptHashTy,
+			scriptInfo: ScriptInfo{
+				PkScriptClass:  ScriptHashTy,
 				NumInputs:      3,
 				ExpectedInputs: 3, // nonstandard p2sh.
 				SigOps:         1,
@@ -436,8 +409,8 @@ func TestCalcScriptInfo(t *testing.T) {
 			pkScript: "HASH160 DATA_20 0xfe441065b6532231de2fac56" +
 				"3152205ec4f59c74 EQUAL",
 			bip16: true,
-			scriptInfo: txscript.ScriptInfo{
-				PkScriptClass:  txscript.ScriptHashTy,
+			scriptInfo: ScriptInfo{
+				PkScriptClass:  ScriptHashTy,
 				NumInputs:      3,
 				ExpectedInputs: -1, // nonstandard p2sh.
 				SigOps:         0,
@@ -456,8 +429,8 @@ func TestCalcScriptInfo(t *testing.T) {
 				"5060708090a0b0c0d0e0f101112131415161718191a1" +
 				"b1c1d1e1f2021 3 CHECKMULTISIG",
 			bip16: true,
-			scriptInfo: txscript.ScriptInfo{
-				PkScriptClass:  txscript.MultiSigTy,
+			scriptInfo: ScriptInfo{
+				PkScriptClass:  MultiSigTy,
 				NumInputs:      4,
 				ExpectedInputs: 3,
 				SigOps:         3,
@@ -468,8 +441,7 @@ func TestCalcScriptInfo(t *testing.T) {
 	for _, test := range tests {
 		sigScript := mustParseShortForm(test.sigScript)
 		pkScript := mustParseShortForm(test.pkScript)
-		si, err := txscript.CalcScriptInfo(sigScript, pkScript,
-			test.bip16)
+		si, err := CalcScriptInfo(sigScript, pkScript, test.bip16)
 		if err != nil {
 			if err != test.scriptInfoErr {
 				t.Errorf("scriptinfo test \"%s\": got \"%v\""+
@@ -541,8 +513,9 @@ func TestPayToAddrScript(t *testing.T) {
 	t.Parallel()
 
 	// 1MirQ9bwyQcGVJPwKUgapu5ouK2E2Ey4gX
-	p2pkhMain, err := dcrutil.NewAddressPubKeyHash(decodeHex("e34cce70c863"+
-		"73273efcc54ce7d2a491bb4a0e84"), &chaincfg.MainNetParams, secp)
+	p2pkhMain, err := dcrutil.NewAddressPubKeyHash(hexToBytes("e34cce70c86"+
+		"373273efcc54ce7d2a491bb4a0e84"), &chaincfg.MainNetParams,
+		chainec.ECTypeSecp256k1)
 	if err != nil {
 		t.Errorf("Unable to create public key hash address: %v", err)
 		return
@@ -550,24 +523,24 @@ func TestPayToAddrScript(t *testing.T) {
 
 	// Taken from transaction:
 	// b0539a45de13b3e0403909b8bd1a555b8cbe45fd4e3f3fda76f3a5f52835c29d
-	p2shMain, _ := dcrutil.NewAddressScriptHashFromHash(decodeHex("e8c300"+
-		"c87986efa84c37c0519929019ef86eb5b4"), &chaincfg.MainNetParams)
+	p2shMain, _ := dcrutil.NewAddressScriptHashFromHash(hexToBytes("e8c30"+
+		"0c87986efa84c37c0519929019ef86eb5b4"), &chaincfg.MainNetParams)
 	if err != nil {
 		t.Errorf("Unable to create script hash address: %v", err)
 		return
 	}
 
 	//  mainnet p2pk 13CG6SJ3yHUXo4Cr2RY4THLLJrNFuG3gUg
-	p2pkCompressedMain, err := dcrutil.NewAddressSecpPubKey(decodeHex("02192d74"+
-		"d0cb94344c9569c2e77901573d8d7903c3ebec3a957724895dca52c6b4"),
+	p2pkCompressedMain, err := dcrutil.NewAddressSecpPubKey(hexToBytes("02192d7"+
+		"4d0cb94344c9569c2e77901573d8d7903c3ebec3a957724895dca52c6b4"),
 		&chaincfg.MainNetParams)
 	if err != nil {
 		t.Errorf("Unable to create pubkey address (compressed): %v",
 			err)
 		return
 	}
-	p2pkCompressed2Main, err := dcrutil.NewAddressSecpPubKey(decodeHex("03b0bd"+
-		"634234abbb1ba1e986e884185c61cf43e001f9137f23c2c409273eb16e65"),
+	p2pkCompressed2Main, err := dcrutil.NewAddressSecpPubKey(hexToBytes("03b0b"+
+		"d634234abbb1ba1e986e884185c61cf43e001f9137f23c2c409273eb16e65"),
 		&chaincfg.MainNetParams)
 	if err != nil {
 		t.Errorf("Unable to create pubkey address (compressed 2): %v",
@@ -575,7 +548,7 @@ func TestPayToAddrScript(t *testing.T) {
 		return
 	}
 
-	p2pkUncompressedMain := newAddressPubKey(decodeHex("0411db" +
+	p2pkUncompressedMain := newAddressPubKey(hexToBytes("0411db" +
 		"93e1dcdb8a016b49840f8c53bc1eb68a382e97b1482ecad7b148a6909a5cb2" +
 		"e0eaddfb84ccf9744464f82e160bfa9b8b64f9d4c03f999b8643f656b412a3"))
 
@@ -623,17 +596,17 @@ func TestPayToAddrScript(t *testing.T) {
 		},
 
 		// Supported address types with nil pointers.
-		{(*dcrutil.AddressPubKeyHash)(nil), "", txscript.ErrUnsupportedAddress},
-		{(*dcrutil.AddressScriptHash)(nil), "", txscript.ErrUnsupportedAddress},
-		{(*dcrutil.AddressSecpPubKey)(nil), "", txscript.ErrUnsupportedAddress},
+		{(*dcrutil.AddressPubKeyHash)(nil), "", ErrUnsupportedAddress},
+		{(*dcrutil.AddressScriptHash)(nil), "", ErrUnsupportedAddress},
+		{(*dcrutil.AddressSecpPubKey)(nil), "", ErrUnsupportedAddress},
 
 		// Unsupported address type.
-		{&bogusAddress{}, "", txscript.ErrUnsupportedAddress},
+		{&bogusAddress{}, "", ErrUnsupportedAddress},
 	}
 
 	t.Logf("Running %d tests", len(tests))
 	for i, test := range tests {
-		pkScript, err := txscript.PayToAddrScript(test.in)
+		pkScript, err := PayToAddrScript(test.in)
 		if err != test.err {
 			t.Errorf("PayToAddrScript #%d unexpected error - "+
 				"got %v, want %v", i, err, test.err)
@@ -655,16 +628,16 @@ func TestMultiSigScript(t *testing.T) {
 	t.Parallel()
 
 	//  mainnet p2pk 13CG6SJ3yHUXo4Cr2RY4THLLJrNFuG3gUg
-	p2pkCompressedMain, err := dcrutil.NewAddressSecpPubKey(decodeHex("02192d7"+
-		"4d0cb94344c9569c2e77901573d8d7903c3ebec3a957724895dca52c6b4"),
+	p2pkCompressedMain, err := dcrutil.NewAddressSecpPubKey(hexToBytes("02192d"+
+		"74d0cb94344c9569c2e77901573d8d7903c3ebec3a957724895dca52c6b4"),
 		&chaincfg.MainNetParams)
 	if err != nil {
 		t.Errorf("Unable to create pubkey address (compressed): %v",
 			err)
 		return
 	}
-	p2pkCompressed2Main, err := dcrutil.NewAddressSecpPubKey(decodeHex("03b0bd"+
-		"634234abbb1ba1e986e884185c61cf43e001f9137f23c2c409273eb16e65"),
+	p2pkCompressed2Main, err := dcrutil.NewAddressSecpPubKey(hexToBytes("03b0b"+
+		"d634234abbb1ba1e986e884185c61cf43e001f9137f23c2c409273eb16e65"),
 		&chaincfg.MainNetParams)
 	if err != nil {
 		t.Errorf("Unable to create pubkey address (compressed 2): %v",
@@ -672,7 +645,7 @@ func TestMultiSigScript(t *testing.T) {
 		return
 	}
 
-	p2pkUncompressedMain := newAddressPubKey(decodeHex("0411d" +
+	p2pkUncompressedMain := newAddressPubKey(hexToBytes("0411d" +
 		"b93e1dcdb8a016b49840f8c53bc1eb68a382e97b1482ecad7b148a6909a5c" +
 		"b2e0eaddfb84ccf9744464f82e160bfa9b8b64f9d4c03f999b8643f656b41" +
 		"2a3"))
@@ -714,7 +687,7 @@ func TestMultiSigScript(t *testing.T) {
 			},
 			3,
 			"",
-			txscript.ErrBadNumRequired,
+			ErrBadNumRequired,
 		},
 		{
 			// By default compressed pubkeys are used in Decred.
@@ -732,14 +705,13 @@ func TestMultiSigScript(t *testing.T) {
 			},
 			2,
 			"",
-			txscript.ErrBadNumRequired,
+			ErrBadNumRequired,
 		},
 	}
 
 	t.Logf("Running %d tests", len(tests))
 	for i, test := range tests {
-		script, err := txscript.MultiSigScript(test.keys,
-			test.nrequired)
+		script, err := MultiSigScript(test.keys, test.nrequired)
 		if err != test.err {
 			t.Errorf("MultiSigScript #%d unexpected error - "+
 				"got %v, want %v", i, err, test.err)
@@ -769,14 +741,14 @@ func TestCalcMultiSigStats(t *testing.T) {
 			name: "short script",
 			script: "0x046708afdb0fe5548271967f1a67130b7105cd6a828" +
 				"e03909a67962e0ea1f61d",
-			err: txscript.ErrStackShortScript,
+			err: ErrStackShortScript,
 		},
 		{
 			name: "stack underflow",
 			script: "RETURN DATA_41 0x046708afdb0fe5548271967f1a" +
 				"67130b7105cd6a828e03909a67962e0ea1f61deb649f6" +
 				"bc3f4cef308",
-			err: txscript.ErrStackUnderflow,
+			err: ErrStackUnderflow,
 		},
 		{
 			name: "multisig script",
@@ -792,7 +764,7 @@ func TestCalcMultiSigStats(t *testing.T) {
 
 	for i, test := range tests {
 		script := mustParseShortForm(test.script)
-		if _, _, err := txscript.CalcMultiSigStats(script); err != test.err {
+		if _, _, err := CalcMultiSigStats(script); err != test.err {
 			t.Errorf("CalcMultiSigStats #%d (%s) unexpected "+
 				"error\ngot: %v\nwant: %v", i, test.name, err,
 				test.err)
@@ -807,21 +779,21 @@ func TestCalcMultiSigStats(t *testing.T) {
 var scriptClassTests = []struct {
 	name   string
 	script string
-	class  txscript.ScriptClass
+	class  ScriptClass
 }{
 	{
 		name: "Pay Pubkey",
 		script: "DATA_65 0x0411db93e1dcdb8a016b49840f8c53bc1eb68a382e" +
 			"97b1482ecad7b148a6909a5cb2e0eaddfb84ccf9744464f82e16" +
 			"0bfa9b8b64f9d4c03f999b8643f656b412a3 CHECKSIG",
-		class: txscript.PubKeyTy,
+		class: PubKeyTy,
 	},
 	// tx 599e47a8114fe098103663029548811d2651991b62397e057f0c863c2bc9f9ea
 	{
 		name: "Pay PubkeyHash",
 		script: "DUP HASH160 DATA_20 0x660d4ef3a743e3e696ad990364e555" +
 			"c271ad504b EQUALVERIFY CHECKSIG",
-		class: txscript.PubKeyHashTy,
+		class: PubKeyHashTy,
 	},
 	// part of tx 6d36bc17e947ce00bb6f12f8e7a56a1585c5a36188ffa2b05e10b4743273a74b
 	// codeseparator parts have been elided. (bitcoin core's checks for
@@ -830,45 +802,45 @@ var scriptClassTests = []struct {
 		name: "multisig",
 		script: "1 DATA_33 0x0232abdc893e7f0631364d7fd01cb33d24da4" +
 			"5329a00357b3a7886211ab414d55a 1 CHECKMULTISIG",
-		class: txscript.MultiSigTy,
+		class: MultiSigTy,
 	},
 	// tx e5779b9e78f9650debc2893fd9636d827b26b4ddfa6a8172fe8708c924f5c39d
 	{
 		name: "P2SH",
 		script: "HASH160 DATA_20 0x433ec2ac1ffa1b7b7d027f564529c57197f" +
 			"9ae88 EQUAL",
-		class: txscript.ScriptHashTy,
+		class: ScriptHashTy,
 	},
 
 	{
 		// Nulldata with no data at all.
 		name:   "nulldata no data",
 		script: "RETURN",
-		class:  txscript.NullDataTy,
+		class:  NullDataTy,
 	},
 	{
 		// Nulldata with single zero push.
 		name:   "nulldata zero",
 		script: "RETURN 0",
-		class:  txscript.NullDataTy,
+		class:  NullDataTy,
 	},
 	{
 		// Nulldata with small integer push.
 		name:   "nulldata small int",
 		script: "RETURN 1",
-		class:  txscript.NullDataTy,
+		class:  NullDataTy,
 	},
 	{
 		// Nulldata with max small integer push.
 		name:   "nulldata max small int",
 		script: "RETURN 16",
-		class:  txscript.NullDataTy,
+		class:  NullDataTy,
 	},
 	{
 		// Nulldata with small data push.
 		name:   "nulldata small data",
 		script: "RETURN DATA_8 0x046708afdb0fe554",
-		class:  txscript.NullDataTy,
+		class:  NullDataTy,
 	},
 	{
 		// Canonical nulldata with 60-byte data push.
@@ -876,7 +848,7 @@ var scriptClassTests = []struct {
 		script: "RETURN 0x3c 0x046708afdb0fe5548271967f1a67130b7105cd" +
 			"6a828e03909a67962e0ea1f61deb649f6bc3f4cef3046708afdb" +
 			"0fe5548271967f1a67130b7105cd6a",
-		class: txscript.NullDataTy,
+		class: NullDataTy,
 	},
 	{
 		// Non-canonical nulldata with 60-byte data push.
@@ -884,7 +856,7 @@ var scriptClassTests = []struct {
 		script: "RETURN PUSHDATA1 0x3c 0x046708afdb0fe5548271967f1a67" +
 			"130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef3" +
 			"046708afdb0fe5548271967f1a67130b7105cd6a",
-		class: txscript.NullDataTy,
+		class: NullDataTy,
 	},
 	{
 		// Nulldata with max allowed data to be considered standard.
@@ -893,7 +865,7 @@ var scriptClassTests = []struct {
 			"130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef3" +
 			"046708afdb0fe5548271967f1a67130b7105cd6a828e03909a67" +
 			"962e0ea1f61deb649f6bc3f4cef3",
-		class: txscript.NullDataTy,
+		class: NullDataTy,
 	},
 	{
 		// Nulldata with more than max allowed data to be considered
@@ -905,14 +877,14 @@ var scriptClassTests = []struct {
 			"fdb0fe5548271967f1a67046708afdb0fe5548271967f1a67046708afdb" +
 			"0fe5548271967f1a67046708afdb0fe5548271967f1a67046708afdb0fe" +
 			"5548271967f1a67",
-		class: txscript.NonStandardTy,
+		class: NonStandardTy,
 	},
 	{
 		// Almost nulldata, but add an additional opcode after the data
 		// to make it nonstandard.
 		name:   "almost nulldata",
 		script: "RETURN 4 TRUE",
-		class:  txscript.NonStandardTy,
+		class:  NonStandardTy,
 	},
 
 	// The next few are almost multisig (it is the more complex script type)
@@ -922,13 +894,13 @@ var scriptClassTests = []struct {
 		name: "strange 1",
 		script: "DUP DATA_33 0x0232abdc893e7f0631364d7fd01cb33d24da45" +
 			"329a00357b3a7886211ab414d55a 1 CHECKMULTISIG",
-		class: txscript.NonStandardTy,
+		class: NonStandardTy,
 	},
 	{
 		// Multisig but invalid pubkey.
 		name:   "strange 2",
 		script: "1 1 1 CHECKMULTISIG",
-		class:  txscript.NonStandardTy,
+		class:  NonStandardTy,
 	},
 	{
 		// Multisig but no matching npubkeys opcode.
@@ -937,25 +909,25 @@ var scriptClassTests = []struct {
 			"9a00357b3a7886211ab414d55a DATA_33 0x0232abdc893e7f0" +
 			"631364d7fd01cb33d24da45329a00357b3a7886211ab414d55a " +
 			"CHECKMULTISIG",
-		class: txscript.NonStandardTy,
+		class: NonStandardTy,
 	},
 	{
 		// Multisig but with multisigverify.
 		name: "strange 4",
 		script: "1 DATA_33 0x0232abdc893e7f0631364d7fd01cb33d24da4532" +
 			"9a00357b3a7886211ab414d55a 1 CHECKMULTISIGVERIFY",
-		class: txscript.NonStandardTy,
+		class: NonStandardTy,
 	},
 	{
 		// Multisig but wrong length.
 		name:   "strange 5",
 		script: "1 CHECKMULTISIG",
-		class:  txscript.NonStandardTy,
+		class:  NonStandardTy,
 	},
 	{
 		name:   "doesn't parse",
 		script: "DATA_5 0x01020304",
-		class:  txscript.NonStandardTy,
+		class:  NonStandardTy,
 	},
 	{
 		name: "multisig script with wrong number of pubkeys",
@@ -967,7 +939,7 @@ var scriptClassTests = []struct {
 			"0x02c08f3de8ee2de9be7bd770f4c10eb0" +
 			"d6ff1dd81ee96eedd3a9d4aeaf86695e80 " +
 			"3 CHECKMULTISIG",
-		class: txscript.NonStandardTy,
+		class: NonStandardTy,
 	},
 }
 
@@ -978,7 +950,7 @@ func TestScriptClass(t *testing.T) {
 
 	for _, test := range scriptClassTests {
 		script := mustParseShortForm(test.script)
-		class := txscript.GetScriptClass(txscript.DefaultScriptVersion, script)
+		class := GetScriptClass(DefaultScriptVersion, script)
 		if class != test.class {
 			t.Errorf("%s: expected %s got %s (script %x)", test.name,
 				test.class, class, script)
@@ -994,42 +966,42 @@ func TestStringifyClass(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		class    txscript.ScriptClass
+		class    ScriptClass
 		stringed string
 	}{
 		{
 			name:     "nonstandardty",
-			class:    txscript.NonStandardTy,
+			class:    NonStandardTy,
 			stringed: "nonstandard",
 		},
 		{
 			name:     "pubkey",
-			class:    txscript.PubKeyTy,
+			class:    PubKeyTy,
 			stringed: "pubkey",
 		},
 		{
 			name:     "pubkeyhash",
-			class:    txscript.PubKeyHashTy,
+			class:    PubKeyHashTy,
 			stringed: "pubkeyhash",
 		},
 		{
 			name:     "scripthash",
-			class:    txscript.ScriptHashTy,
+			class:    ScriptHashTy,
 			stringed: "scripthash",
 		},
 		{
 			name:     "multisigty",
-			class:    txscript.MultiSigTy,
+			class:    MultiSigTy,
 			stringed: "multisig",
 		},
 		{
 			name:     "nulldataty",
-			class:    txscript.NullDataTy,
+			class:    NullDataTy,
 			stringed: "nulldata",
 		},
 		{
 			name:     "broken",
-			class:    txscript.ScriptClass(255),
+			class:    ScriptClass(255),
 			stringed: "Invalid",
 		},
 	}
@@ -1050,21 +1022,21 @@ func TestGenerateProvablyPruneableOut(t *testing.T) {
 		data     []byte
 		expected []byte
 		err      error
-		class    txscript.ScriptClass
+		class    ScriptClass
 	}{
 		{
 			name:     "small int",
 			data:     hexToBytes("01"),
 			expected: mustParseShortForm("RETURN 1"),
 			err:      nil,
-			class:    txscript.NullDataTy,
+			class:    NullDataTy,
 		},
 		{
 			name:     "max small int",
 			data:     hexToBytes("10"),
 			expected: mustParseShortForm("RETURN 16"),
 			err:      nil,
-			class:    txscript.NullDataTy,
+			class:    NullDataTy,
 		},
 		{
 			name: "data of size before OP_PUSHDATA1 is needed",
@@ -1073,7 +1045,7 @@ func TestGenerateProvablyPruneableOut(t *testing.T) {
 			expected: mustParseShortForm("RETURN 0x18 0x01020304" +
 				"05060708090a0b0c0d0e0f101112131415161718"),
 			err:   nil,
-			class: txscript.NullDataTy,
+			class: NullDataTy,
 		},
 		{
 			name: "just right",
@@ -1100,7 +1072,7 @@ func TestGenerateProvablyPruneableOut(t *testing.T) {
 				"b4c4d4e4f202122232425262728292a2b2c2d2e2f30313233" +
 				"3435363738393a3b3c3d3e"),
 			err:   nil,
-			class: txscript.NullDataTy,
+			class: NullDataTy,
 		},
 		{
 			name: "too big",
@@ -1114,13 +1086,13 @@ func TestGenerateProvablyPruneableOut(t *testing.T) {
 				"4a4b4c4d4e4f202122232425262728292a2b2c2d2e2f303132333435363738393" +
 				"a3b3c3d3e3f3f"),
 			expected: nil,
-			err:      txscript.ErrStackLongScript,
-			class:    txscript.NonStandardTy,
+			err:      ErrStackLongScript,
+			class:    NonStandardTy,
 		},
 	}
 
 	for i, test := range tests {
-		script, err := txscript.GenerateProvablyPruneableOut(test.data)
+		script, err := GenerateProvablyPruneableOut(test.data)
 		if err != test.err {
 			t.Errorf("GenerateProvablyPruneableOut: #%d (%s) unexpected error: "+
 				"got %v, want %v", i, test.name, err, test.err)
@@ -1136,7 +1108,7 @@ func TestGenerateProvablyPruneableOut(t *testing.T) {
 		}
 
 		// Check that the script has the correct type.
-		scriptType := txscript.GetScriptClass(txscript.DefaultScriptVersion, script)
+		scriptType := GetScriptClass(DefaultScriptVersion, script)
 		if scriptType != test.class {
 			t.Errorf("GetScriptClass: #%d (%s) wrong result -- "+
 				"got: %v, want: %v", i, test.name, scriptType,
